update : method overriding at impala_ds
